moved styling into assets folder
