+++ src/assets/styles/NewGame.css
