+++ src/components/User.js
@@ -25,7 +25,7 @@ function User(props) {
 
     return (
         <section className="userbar">
-            <img src={props.img ? props.img : "https://placekitten.com/50" } />
+            <img src={props.profile_pic ? props.profile_pic : "https://placekitten.com/50" } />
             <section className="userinfo">
                 <h3>{props.username}</h3>
                 <p>{props.title}</p>
