+++ src/pages/HostGame.js
@@ -2,7 +2,7 @@
 import React, { useEffect, useState } from "react";
 import Lobby from "../components/Lobby";
 import { v4 as uuidv4 } from 'uuid';
-import "../styles/NewGame.css";
+import "../assets/styles/NewGame.css";
 
 
 export default function HostGame({socket}) {
