+++ src/pages/Search.js
@@ -1,6 +1,7 @@
 import { useState, useEffect } from "react";
-import API from "../utils/API"
-import User from "../components/User"
+import API from "../utils/API";
+import User from "../components/User";
+import "../assets/styles/Search.css";
 
 function Search() {
     const [searchData, setSearchData] = useState([]);
@@ -27,7 +28,7 @@ function Search() {
     }
 
     return (
-        <section className="page">
+        <section className="page container d-flex align-items-center justify-content-center">
             <section className="search subpage">
                 <form>
                     <input 
@@ -35,6 +36,7 @@ function Search() {
                         type="text"
                         value={searchField}
                         onChange={handleInputChange}
+                        placeholder="search for someone by username!"
                     />
                     <button
                         onClick={handleSubmit}
@@ -42,7 +44,7 @@ function Search() {
                 </form>
                 {searchData.map(user=>{
                     // TODO: Add friend status to this, as well
-                    return <a href={`/profile/${user.username}`} key={user.id}><User username={user.username} title={user.current_title}/></a>
+                    return <a href={`/profile/${user.username}`} key={user.id}><User pic={user.profile_pic} username={user.username} title={user.current_title}/></a>
                 })}
             </section>
         </section>
